fix: resolve comments in pl
