+++ front/src/modules/settings/components/SettingsNavbar.tsx
@@ -12,9 +12,9 @@ import {
   IconUserCircle,
   IconUsers,
 } from '@/ui/display/icon/index';
-import NavItem from '@/ui/navigation/desktop-navbar/components/NavItem';
-import NavTitle from '@/ui/navigation/desktop-navbar/components/NavTitle';
-import SubMenuNavbar from '@/ui/navigation/desktop-navbar/components/SubMenuNavbar';
+import NavItem from '@/ui/navigation/navbar/desktop-navbar/components/NavItem';
+import NavTitle from '@/ui/navigation/navbar/desktop-navbar/components/NavTitle';
+import SubMenuNavbar from '@/ui/navigation/navbar/desktop-navbar/components/SubMenuNavbar';
 
 export const SettingsNavbar = () => {
   const navigate = useNavigate();
