+++ front/src/modules/ui/navigation/navbar/desktop-navbar/components/NavCollapseButton.tsx
