+++ front/src/modules/ui/navigation/navbar/desktop-navbar/components/NavItemsContainer.tsx
