+++ front/src/modules/ui/navigation/navbar/desktop-navbar/components/NavWorkspaceButton.tsx
