+++ front/src/modules/ui/navigation/navbar/desktop-navbar/components/__stories__/MainNavbar.stories.tsx
