+++ front/src/modules/ui/navigation/navbar/desktop-navbar/constants/index.ts
